Add LanguageProficiencies to StudentSupportCallResponse

+++ src/main/resources/db/migration/V216__Add_LanguageProficiencies_to_StudentSupportCallResponses.sql
@@ -1,2 +1,2 @@
 ALTER TABLE `StudentSupportCallResponses` ADD COLUMN `CollectLanguageProficiencies` TINYINT(1) NOT NULL DEFAULT '0';
-ALTER TABLE `StudentSupportCallResponses` ADD COLUMN `LanguageProficiency` INT(1) NULL;
+ALTER TABLE `StudentSupportCallResponses` ADD COLUMN `LanguageProficiency` TINYINT(1) NULL;
